+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-# resourcet-effectful-0.1 (2022-07-14)
+# resourcet-effectful-1.0.1.0 (2023-??-??)
+* Add `allocateEff`, `allocateEff_` and `registerEff`.
+
+# resourcet-effectful-1.0.0.0 (2022-07-14)
 * Initial release.